requirements.txt
certifi
chardet
defusedxml
dj-database-url
Django
asgiref~=3.2.10
django-allauth
django-cors-headers
django-rest-auth
django-storages
djangorestframework
dropbox
idna
oauthlib
psycopg2
PyMySQL
python3-openid
pytz
requests
requests-oauthlib
six
sqlparse
toml
Unidecode
urllib3
whitenoise
google-cloud-tasks
